issue#25: backend be able to create slugs for duplicate
